refactor span controller and routes
