--- service/web-server/src/routes/span.js
@@ -1,13 +1,0 @@
-/**
- * /routes/span
- */
-
-const restify = require('restify');
-
-const controller = require('../controllers/span');
-
-const router = restify.createServer()
-
-router.post('/query', controller);
-
-module.exports = router;
